+++ src/app/locales/en/utils/user-actions/leave-action.js
@@ -1,0 +1,5 @@
+export default {
+  title: 'Leave',
+  successNotificationText: '{{readableModelName}} left successfully',
+  failureNotificationActionName: 'leaving {{readableModelName}}',
+};