+++ CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
 ### Latest changes
 
+* VFS-11134 Added information about membership to leaving modal
 * VFS-10939 Fixed copying from provider popover in firefox
 * VFS-11202 Removed Embedded Ceph cluster configuration
 * VFS-11225 Added failure threshold to atm lane